Change some tmp array types to smaller ones

+++ Source/astcenc_block_sizes2.cpp
@@ -117,8 +117,10 @@ static int decode_block_mode_2d(
 	int weight_count = N * M * (D + 1);
 	int qmode = (base_quant_mode - 2) + 6 * H;
 
-	int weightbits = compute_ise_bitcount(weight_count, (quantization_method) qmode);
-	if (weight_count > MAX_WEIGHTS_PER_BLOCK || weightbits < MIN_WEIGHT_BITS_PER_BLOCK || weightbits > MAX_WEIGHT_BITS_PER_BLOCK)
+	int weightbits = compute_ise_bitcount(weight_count, (quantization_method)qmode);
+	if (weight_count > MAX_WEIGHTS_PER_BLOCK ||
+	    weightbits < MIN_WEIGHT_BITS_PER_BLOCK ||
+	    weightbits > MAX_WEIGHT_BITS_PER_BLOCK)
 	{
 		return 0;
 	}
@@ -211,7 +213,7 @@ static int decode_block_mode_3d(
 	int weight_count = N * M * Q * (D + 1);
 	int qmode = (base_quant_mode - 2) + 6 * H;
 
-	int weightbits = compute_ise_bitcount(weight_count, (quantization_method) qmode);
+	int weightbits = compute_ise_bitcount(weight_count, (quantization_method)qmode);
 	if (weight_count > MAX_WEIGHTS_PER_BLOCK ||
 	    weightbits < MIN_WEIGHT_BITS_PER_BLOCK ||
 	    weightbits > MAX_WEIGHT_BITS_PER_BLOCK)
@@ -237,12 +239,12 @@ static void initialize_decimation_table_2d(
 	int texels_per_block = xdim * ydim;
 	int weights_per_block = x_weights * y_weights;
 
-	int weightcount_of_texel[MAX_TEXELS_PER_BLOCK];
-	int grid_weights_of_texel[MAX_TEXELS_PER_BLOCK][4];
-	int weights_of_texel[MAX_TEXELS_PER_BLOCK][4];
+	uint8_t weightcount_of_texel[MAX_TEXELS_PER_BLOCK];
+	uint8_t grid_weights_of_texel[MAX_TEXELS_PER_BLOCK][4];
+	uint8_t weights_of_texel[MAX_TEXELS_PER_BLOCK][4];
 
-	int texelcount_of_weight[MAX_WEIGHTS_PER_BLOCK];
-	int texels_of_weight[MAX_WEIGHTS_PER_BLOCK][MAX_TEXELS_PER_BLOCK];
+	uint8_t texelcount_of_weight[MAX_WEIGHTS_PER_BLOCK];
+	uint8_t texels_of_weight[MAX_WEIGHTS_PER_BLOCK][MAX_TEXELS_PER_BLOCK];
 	int texelweights_of_weight[MAX_WEIGHTS_PER_BLOCK][MAX_TEXELS_PER_BLOCK];
 
 	for (int i = 0; i < weights_per_block; i++)
@@ -302,8 +304,7 @@ static void initialize_decimation_table_2d(
 	{
 		dt->texel_weight_count[i] = weightcount_of_texel[i];
 
-		// ensure that all 4 entries are actually initialized.
-		// This allows a branch-free implementation of compute_value_of_texel_flt()
+		// Init all 4 entries so we can rely on zeros for vectorization
 		for (int j = 0; j < 4; j++)
 		{
 			dt->texel_weights_int_t4[i][j] = 0;
@@ -317,13 +318,12 @@ static void initialize_decimation_table_2d(
 
 		for (int j = 0; j < weightcount_of_texel[i]; j++)
 		{
-			// TODO: Why the uint8_t casts? Can we make these smaller?
-			dt->texel_weights_int_t4[i][j] = (uint8_t)weights_of_texel[i][j];
+			dt->texel_weights_int_t4[i][j] = weights_of_texel[i][j];
 			dt->texel_weights_float_t4[i][j] = ((float)weights_of_texel[i][j]) * (1.0f / TEXEL_WEIGHT_SUM);
-			dt->texel_weights_t4[i][j] = (uint8_t)grid_weights_of_texel[i][j];
+			dt->texel_weights_t4[i][j] = grid_weights_of_texel[i][j];
 
 			dt->texel_weights_float_4t[j][i] = ((float)weights_of_texel[i][j]) * (1.0f / TEXEL_WEIGHT_SUM);
-			dt->texel_weights_4t[j][i] = (uint8_t)grid_weights_of_texel[i][j];
+			dt->texel_weights_4t[j][i] = grid_weights_of_texel[i][j];
 		}
 	}
 
@@ -333,9 +333,9 @@ static void initialize_decimation_table_2d(
 
 		for (int j = 0; j < texelcount_of_weight[i]; j++)
 		{
-			int texel = texels_of_weight[i][j];
-			dt->weight_texel[i][j] = (uint8_t)texel;
-			dt->weights_int[i][j] = (uint8_t)texelweights_of_weight[i][j];
+			uint8_t texel = texels_of_weight[i][j];
+			dt->weight_texel[i][j] = texel;
+			dt->weights_int[i][j] = texelweights_of_weight[i][j];
 			dt->weights_flt[i][j] = (float)texelweights_of_weight[i][j];
 
 			// perform a layer of array unrolling. An aspect of this unrolling is that
@@ -344,23 +344,23 @@ static void initialize_decimation_table_2d(
 			int swap_idx = -1;
 			for (int k = 0; k < 4; k++)
 			{
-				int dttw = dt->texel_weights_t4[texel][k];
+				uint8_t dttw = dt->texel_weights_t4[texel][k];
 				float dttwf = dt->texel_weights_float_t4[texel][k];
 				if (dttw == i && dttwf != 0.0f)
 				{
 					swap_idx = k;
 				}
-				dt->texel_weights_texel[i][j][k] = (uint8_t)dttw;
+				dt->texel_weights_texel[i][j][k] = dttw;
 				dt->texel_weights_float_texel[i][j][k] = dttwf;
 			}
 
 			if (swap_idx != 0)
 			{
-				int vi = dt->texel_weights_texel[i][j][0];
+				uint8_t vi = dt->texel_weights_texel[i][j][0];
 				float vf = dt->texel_weights_float_texel[i][j][0];
 				dt->texel_weights_texel[i][j][0] = dt->texel_weights_texel[i][j][swap_idx];
 				dt->texel_weights_float_texel[i][j][0] = dt->texel_weights_float_texel[i][j][swap_idx];
-				dt->texel_weights_texel[i][j][swap_idx] = (uint8_t)vi;
+				dt->texel_weights_texel[i][j][swap_idx] = vi;
 				dt->texel_weights_float_texel[i][j][swap_idx] = vf;
 			}
 		}
@@ -520,8 +520,7 @@ static void initialize_decimation_table_3d(
 	{
 		dt->texel_weight_count[i] = weightcount_of_texel[i];
 
-		// ensure that all 4 entries are actually initialized.
-		// This allows a branch-free implementation of compute_value_of_texel_flt()
+		// Init all 4 entries so we can rely on zeros for vectorization
 		for (int j = 0; j < 4; j++)
 		{
 			dt->texel_weights_int_t4[i][j] = 0;
